+++ CHANGELOG.md
@@ -65,6 +65,7 @@ Changes can also be flagged with a GitHub label for tracking purposes. The URL o
 - Fixed details requirements in AWS SES setup [#6047](https://github.com/ethyca/fides/pull/6047)
 - Addressed some performance issues with Experience configuration previews [#6055](https://github.com/ethyca/fides/pull/6055)
 - Fixed icon sizing in request manager table [#6079](https://github.com/ethyca/fides/pull/6079)
+- Fixed GCP SQL connection to support ip_type [#6065](https://github.com/ethyca/fides/pull/6065)
 
 ## [2.59.2](https://github.com/ethyca/fides/compare/2.59.1...2.59.2)